pyproject.toml
[project]
name = "sound-sozai"
version = "0.1.0"
description = "Audio file transcription and splitting tool using Whisper"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "openai-whisper>=20231117",
    "pydub>=0.25.1",
    "tqdm>=4.66.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }
